+++ pkg/services/container/server_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/neorpc/result"
 	"github.com/nspcc-dev/neo-go/pkg/smartcontract/trigger"
 	containerSvc "github.com/nspcc-dev/neofs-node/pkg/services/container"
+	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
 	"github.com/nspcc-dev/neofs-sdk-go/container"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
@@ -78,15 +79,23 @@ func (x testNodeState) CurrentEpoch() uint64 { return x.epoch }
 
 type testFSChain struct {
 	testNodeState
-	cnr container.Container
+	getErr error
+	cnrs   map[cid.ID]container.Container
 }
 
 func (testFSChain) Put(container.Container, []byte, []byte, *session.Container) (cid.ID, error) {
 	return cid.ID{}, errors.New("unimplemented")
 }
 
-func (x testFSChain) Get(cid.ID) (container.Container, error) {
-	return x.cnr, nil
+func (x testFSChain) Get(id cid.ID) (container.Container, error) {
+	if x.getErr != nil {
+		return container.Container{}, x.getErr
+	}
+	cnr, ok := x.cnrs[id]
+	if !ok {
+		return container.Container{}, apistatus.ErrContainerNotFound
+	}
+	return cnr, nil
 }
 
 func (testFSChain) List(user.ID) ([]cid.ID, error) {
@@ -94,7 +103,7 @@ func (testFSChain) List(user.ID) ([]cid.ID, error) {
 }
 
 func (testFSChain) PutEACL(eacl.Table, []byte, []byte, *session.Container) error {
-	return errors.New("unimplemented")
+	return nil
 }
 
 func (testFSChain) GetEACL(cid.ID) (eacl.Table, error) {
@@ -139,13 +148,16 @@ func makeDeleteRequestWithSessionMessage(t testing.TB, usr usertest.UserSigner, 
 func TestServer_Delete(t *testing.T) {
 	ctx := context.Background()
 	usr := usertest.User()
+	id := cidtest.ID()
 
 	const anyEpoch = 10
 	var cnr container.Container
 	cnr.SetOwner(usr.ID)
 
 	m := &testFSChain{
-		cnr: cnr,
+		cnrs: map[cid.ID]container.Container{
+			id: cnr,
+		},
 	}
 	m.epoch = anyEpoch
 	svc := containerSvc.New(&usr.ECDSAPrivateKey, m, m, m, m)
@@ -241,7 +253,7 @@ func TestServer_Delete(t *testing.T) {
 				otherUsr := usertest.User()
 				require.NoError(t, st.Sign(otherUsr))
 
-				req := makeDeleteRequestWithSession(t, usr, cidtest.ID(), st)
+				req := makeDeleteRequestWithSession(t, usr, id, st)
 				resp, err := svc.Delete(ctx, req)
 				require.NoError(t, err)
 				require.NotNil(t, resp)
@@ -331,8 +343,13 @@ func TestSessionVerb(t *testing.T) {
 	owner := usertest.User()
 	cnrID := cidtest.ID()
 
+	var cnr container.Container
+	cnr.SetOwner(owner.ID)
+
 	var fsChain testFSChain
-	fsChain.cnr.SetOwner(owner.ID)
+	fsChain.cnrs = map[cid.ID]container.Container{
+		cnrID: cnr,
+	}
 
 	s := containerSvc.New(&owner.ECDSAPrivateKey, fsChain, fsChain, fsChain, fsChain)
 
@@ -447,4 +464,194 @@ func TestServer_SetExtendedACL_InvalidRequest(t *testing.T) {
 			},
 		}))
 	})
+}
+
+func TestService_SetExtendedACL_SessionIssuer(t *testing.T) {
+	ctx := context.Background()
+	const currentEpoch = 10
+
+	owner := usertest.User()
+	id := cidtest.ID()
+	var cnr container.Container
+	cnr.SetOwner(owner.ID)
+
+	otherID := cidtest.OtherID(id)
+	var otherCnr container.Container
+	otherCnr.SetOwner(usertest.OtherID(owner.ID))
+
+	var fsChain testFSChain
+	fsChain.epoch = currentEpoch
+	fsChain.cnrs = map[cid.ID]container.Container{
+		id:      cnr,
+		otherID: otherCnr,
+	}
+
+	s := containerSvc.New(&owner.ECDSAPrivateKey, &fsChain, &fsChain, &fsChain, &fsChain)
+
+	eACL := eacltest.Table()
+
+	var st session.Container
+	st.SetID(uuid.New())
+	st.SetExp(fsChain.epoch + 1)
+	st.SetAuthKey(neofscryptotest.Signer().Public())
+	st.ForVerb(session.VerbContainerSetEACL)
+	require.NoError(t, st.Sign(owner))
+
+	req := &protocontainer.SetExtendedACLRequest{
+		Body: &protocontainer.SetExtendedACLRequest_Body{
+			Signature: &refs.SignatureRFC6979{},
+		},
+		MetaHeader: &protosession.RequestMetaHeader{
+			SessionToken: st.ProtoMessage(),
+		},
+	}
+
+	call := func(t *testing.T, id cid.ID, signer usertest.UserSigner) *protocontainer.SetExtendedACLResponse {
+		eACL.SetCID(id)
+		req.Body.Eacl = eACL.ProtoMessage()
+
+		var err error
+		req.VerifyHeader, err = neofscrypto.SignRequestWithBuffer(owner, req, nil)
+		require.NoError(t, err)
+
+		resp, err := s.SetExtendedACL(ctx, req)
+		require.NoError(t, err)
+		require.NotNil(t, resp)
+		require.NoError(t, neofscrypto.VerifyResponseWithBuffer(resp, nil))
+
+		req.VerifyHeader = nil
+
+		resp.VerifyHeader = nil
+		return resp
+	}
+
+	assertResp := func(t *testing.T, resp *protocontainer.SetExtendedACLResponse, st *protostatus.Status) {
+		require.True(t, proto.Equal(resp, &protocontainer.SetExtendedACLResponse{
+			MetaHeader: &protosession.ResponseMetaHeader{
+				Version: version.Current().ProtoMessage(),
+				Epoch:   currentEpoch,
+				Status:  st,
+			},
+		}), resp.GetMetaHeader().GetStatus())
+	}
+
+	assertOK := func(t *testing.T) {
+		resp := call(t, id, owner)
+		assertResp(t, resp, nil)
+	}
+
+	t.Run("unavailable container", func(t *testing.T) {
+		someError := errors.New("some error")
+		fsChain.getErr = someError
+
+		resp := call(t, id, owner)
+
+		assertResp(t, resp, &protostatus.Status{
+			Code:    1024,
+			Message: someError.Error(),
+		})
+
+		fsChain.getErr = nil
+	})
+
+	assertOK(t)
+
+	t.Run("not a container owner", func(t *testing.T) {
+		resp := call(t, otherID, owner)
+
+		assertResp(t, resp, &protostatus.Status{
+			Code:    1024,
+			Message: "session was not issued by the container owner",
+		})
+	})
+
+	assertOK(t)
+}
+
+func TestService_Delete_SessionIssuer(t *testing.T) {
+	ctx := context.Background()
+	const currentEpoch = 10
+
+	owner := usertest.User()
+	id := cidtest.ID()
+	var cnr container.Container
+	cnr.SetOwner(owner.ID)
+
+	otherID := cidtest.OtherID(id)
+	var otherCnr container.Container
+	otherCnr.SetOwner(usertest.OtherID(owner.ID))
+
+	var fsChain testFSChain
+	fsChain.epoch = currentEpoch
+	fsChain.cnrs = map[cid.ID]container.Container{
+		id:      cnr,
+		otherID: otherCnr,
+	}
+
+	s := containerSvc.New(&owner.ECDSAPrivateKey, &fsChain, &fsChain, &fsChain, &fsChain)
+
+	eACL := eacltest.Table()
+
+	var st session.Container
+	st.SetID(uuid.New())
+	st.SetExp(fsChain.epoch + 1)
+	st.SetAuthKey(neofscryptotest.Signer().Public())
+	st.ForVerb(session.VerbContainerDelete)
+	require.NoError(t, st.Sign(owner))
+
+	call := func(t *testing.T, id cid.ID, signer usertest.UserSigner) *protocontainer.DeleteResponse {
+		eACL.SetCID(id)
+
+		req := makeDeleteRequestWithSession(t, owner, id, st)
+
+		resp, err := s.Delete(ctx, req)
+		require.NoError(t, err)
+		require.NotNil(t, resp)
+		require.NoError(t, neofscrypto.VerifyResponseWithBuffer(resp, nil))
+
+		resp.VerifyHeader = nil
+		return resp
+	}
+
+	assertResp := func(t *testing.T, resp *protocontainer.DeleteResponse, st *protostatus.Status) {
+		require.True(t, proto.Equal(resp, &protocontainer.DeleteResponse{
+			MetaHeader: &protosession.ResponseMetaHeader{
+				Version: version.Current().ProtoMessage(),
+				Epoch:   currentEpoch,
+				Status:  st,
+			},
+		}), resp.GetMetaHeader().GetStatus())
+	}
+
+	assertOK := func(t *testing.T) {
+		resp := call(t, id, owner)
+		assertResp(t, resp, nil)
+	}
+
+	t.Run("unavailable container", func(t *testing.T) {
+		someError := errors.New("some error")
+		fsChain.getErr = someError
+
+		resp := call(t, id, owner)
+
+		assertResp(t, resp, &protostatus.Status{
+			Code:    1024,
+			Message: someError.Error(),
+		})
+
+		fsChain.getErr = nil
+	})
+
+	assertOK(t)
+
+	t.Run("not a container owner", func(t *testing.T) {
+		resp := call(t, otherID, owner)
+
+		assertResp(t, resp, &protostatus.Status{
+			Code:    1024,
+			Message: "session was not issued by the container owner",
+		})
+	})
+
+	assertOK(t)
 }